Use a regular int for the tagged pointer class ID
FossilOrigin-Name: a8d453813ca8007658d52532885acd9c13cad3f0a8b5221985b54ad5a1d057ba

+++ src/runtime/amiga-library.m
@@ -147,7 +147,7 @@ extern void glue_objc_hashtable_set(void);
 extern void *glue_objc_hashtable_get(void);
 extern void glue_objc_hashtable_delete(void);
 extern void glue_objc_hashtable_free(void);
-extern int_fast8_t glue_objc_registerTaggedPointerClass(void);
+extern int glue_objc_registerTaggedPointerClass(void);
 extern Class _Nullable glue_object_getTaggedPointerClass(void);
 extern uintptr_t glue_object_getTaggedPointerValue(void);
 extern id _Nullable glue_objc_createTaggedPointer(void);
